+++ internal/tftime/resource_time_offset.go
@@ -232,6 +232,12 @@ func resourceTimeOffset() *schema.Resource {
 				Type:     schema.TypeInt,
 				Computed: true,
 			},
+			"keepers": {
+				Type:     schema.TypeMap,
+				Optional: true,
+				ForceNew: true,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+			},
 			"minute": {
 				Type:     schema.TypeInt,
 				Computed: true,